requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
mysql-connector-python
alembic
python-dotenv
